Fix canisters e2e test flakiness (#6309)
# Motivation

The `{#each}` loop in `frontend/src/lib/pages/Canisters.svelte` used
`canister.canister_id` instead of `canister.canister_id.toText()` as the
key.
This doesn't work because the same principal can be represented by
different instances.
So when the `canistersStore` was repopulated with the same canisters,
the DOM elements were not reused.

With query-and-update this meant that when the update call results were
put in the store, the canister cards were recreated. Because of this,
occasionally the click on the canister card would fail.

# Changes

1. Use `canister.canister_id.toText()` as the key in the `{#each}` loop.

# Tests

1. I was able to reproduce the flakiness by putting the navigation step
in a loop and perform it many times per test run. With this change, I no
longer observe the flakiness.
2. Created a unit test with a child component of the `CanisterCard` mock
out to be able to tell if the canister card was recreated.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary