refactor(utils): extract csv to file generation utilities (#6308)
# Motivation

#6255 introduced a bug in how `neuronId`s are rendered in the CSV file.
Unit tests did not catch this issue because they focused on the
generated data structure rather than the final string that appears in
the CSV.

This first PR separates the reporting utilities into a secondary
utility. This change is necessary because of how Vitest handles spies.
The `generateCsvFileToSave` function would create a closure on
`saveGeneratedCsv`, which would require mocking the entire file. This
separation allows for the creation of new tests that can spy on the
final value included in the CSV.

A follow-up PR will update the existing tests in
`ReportingNeuronsButton` and `ReportingTransactionsButton` to verify the
CSV string instead of the data structure.

A third PR will address the bug.

No logical changes are introduced in this PR.

# Changes

- Extract the logic for saving a string to a CSV into its own utility.
- Move the relevant tests to their own specification file.
- While updating the imports for `ReportingNeuronsButton` and
`ReportingTransactionsButton`, I also sorted them.

# Tests

- Should pass as before

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary