test(reporting): refactor csv export tests to verify string output (#6310)
# Motivation

#6255 introduced a bug in the rendering of neuronIds in the CSV file.
Unit tests failed to catch this issue because they focused on the
generated data structure instead of the final string in the CSV.

This follows up on #6308 by refactoring tests to monitor the new
utility. This change allows for comparison against the final string
rather than the intermediate data structure.

A follow-up PR will address the bug introduced in #6255.

# Changes

- Update `ReportingNeuronsButton` and `ReportingTransactionsButton` to
spy on `saveGeneratedCsv`

# Tests

- Should pass as before

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary