Simplify State/AccountsStore lifecycle by avoiding replacing accounts_db (#6302)
# Motivation

The `AccountsDb::Map(AccountsDbAsMap)` enum value (i.e. representing
accounts as a map in heap memory) is deprecated, and we'd like to remove
this case (and only have the branch of code for the
`AccountsDb::UnboundedStableBTreeMap(AccountsDbAsUnboundedStableBTreeMap<>)`).
However, currently it's impossible, since `AccountsStore::decode` calls
`AccountsDbAsProxy::default()`, which is an instance of the deprecated
type. The `accounts_db` is later
[replaced](https://sourcegraph.com/github.com/dfinity/nns-dapp@395ad2a74c9309b2f6b943bce3964921904d6e36/-/blob/rs/backend/src/state.rs?L147)
so that the production code indeed uses the `StableBTreeMap`. Therefore,
this PR aims at removing the default-then-replace approach, by
initializing the `AccountsDb::UnboundedStableBTreeMap` directly.


# Changes

- Inline and remove `State::save_heap_to_managed_memory` and
`State::recover_heap_from_managed_memory`, since the logic is simple (so
the indirection is no longer necessary)
- Move some of the logic from `State::new_restored` into `State::decode`
and remove the `default`/`replace_accounts_db` logic (which is the main
goal of the PR)
- Move the `accounts_store` initialization logic from `State::new` to
`AccountsStore::new`, to keep StableBTreeMap initialization logic close
to each other

# Tests

N/A since there are no functionality changes. The test in
`rs/backend/src/state/tests.rs` should (still) cover the code being
changed in this PR.

# Related PRs

[Previous](https://github.com/dfinity/nns-dapp/pull/6301) |
[Next](https://github.com/dfinity/nns-dapp/pull/6311)