feat(portfolio): add authenticated wallet navigation from top token positions (#6335)
# Motivation

We want signed-in users to be able to navigate to their wallets from
their top positions on the Portfolio page.

Note that the Tokens page also allows users to access the wallet even if
they are not logged in. We want to avoid this behavior on the Portfolio
page to maintain consistency with the Staked Tokens card (nextPr), as we
do not allow this on the Staking page.


https://github.com/user-attachments/assets/fd25259d-64b7-4540-ba00-7731d25fa68a

A follow-up PR will add the behavior to the staked tokens card.

# Changes

- Make `HeldTokensCard` rows clickable if user is signed-in.

# Tests

- Added unit tests to verify that the correct tag is rendered: a `div`
when not signed in and an `a` when signed in.
- Manually tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network/portfolio/)

# Todos

- [ ] Add entry to changelog (if necessary).