test: mock ResizeObserver (#6337)
# Motivation

We have to disable the transition for testing purposes, as Svelte v5
introduces breaking changes in how it handles transitions. It now uses
the Web Animations API, which is not available in JSDOM or Happy-DOM.

While testing this new setup against `main`, we discovered that
disabling animations causes Svelte to use the `ResizeObserver`, which is
also an API not currently available in JSDOM or Happy-DOM (see this [CI
job](https://github.com/dfinity/nns-dapp/actions/runs/13131513035/job/36637485021)).

That is why we are mocking this API.

Worth noting, we would need to mock it anyway to integrate Svelte v5.

# References

- https://github.com/dfinity/gix-components/pull/573
- https://github.com/dfinity/gix-components/pull/573

# Changes

<!-- List the changes that have been developed -->

# Tests

After mocking the `ResizeObserver`, the CI passed for the branch (see PR
#6336) that was use to assert this new feature.

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>