feat(navigation): implement Neuron's page entry point tracking  (#6343)
# Motivation

Following a similar approach to #6321, the neurons page will include
multiple return pages based on how it was accessed.

Currently, there is only one entry point: the Staking page, which is
accessed by clicking on any `ProjectsTable` row.

A new entry point will be added in #6342, so we need to know the origin
to determine where to navigate back.

A follow-up PR will update the Neuron's layout to utilize this store.

# Changes

- A new derived store has been created to determine the return
destination from the Neuron's page.

# Tests

- Unit tests have been implemented for this new derived store.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary