feat(portfolio): add authenticated neurons navigation from top projects (#6342)
# Motivation

We want signed-in users to navigate to their wallets from their top
positions on the Portfolio page.

This follows up on #6335. It adds navigation to the rows of the staked
tokens table. Similar to the held tokens card, navigation will only be
available to signed-in users.


https://github.com/user-attachments/assets/1e7e52e3-f24b-439d-9fb3-f171001c4ed6

# Changes

- Make `StakedTokensCard` rows clickable if user is signed-in.

# Tests

- Added unit tests to verify that the correct tag is rendered: a `div`
when not signed in and an `a` when signed in, with the appropriate href
value passed.
- Manually tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network/portfolio/)

# Todos

- [x] Add entry to changelog (if necessary).