Util hasEnoughDissolveDelayToVote (#6353)
# Motivation  

In today's implementation, neurons can be unlocked but inactive for so
long that the "confirm following" warning appears. This is undesired, so
the followee confirmation warning shouldn't be shown if the neuron
cannot vote.
To determine in a unified way whether an NNS neuron has enough dissolve
delay to vote, we introduce a utility function. It will be used to
filter out neurons for which the confirm following notifications should
not be displayed.

# Changes  

- Added a new utility function: `hasEnoughDissolveDelayToVote`.  
- Applied it where neuron voting eligibility is determined.  

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.