Remove MultiPartTransactionProcessor (#6357)
# Motivation

https://github.com/dfinity/nns-dapp/pull/6243 removed the
`multi_part_transactions_processor` field from the `AccountsStore` but
we could not yet remove the `MultiPartTransactionProcessor` because the
previous version of the nns-dapp canister would try to decode a
`MultiPartTransactionProcessor` from stable memory so the current
version still needed to encode one to stable memory in case we would
decide to revert to the previous canister.

Now that's we've had a release, it's safe to stop encoding
`MultiPartTransactionProcessor` to stable memory so we can remove the
type entirely.

# Changes

1. Encode just an empty tuple to the MultiPartTransactionProcessor's
slot in stable memory.
2. Remove `MultiPartTransactionProcessor's`.

# Tests

1. Downgrade-upgrade test still passes.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary