Await async expects (#6383)
# Motivation

In noticed output like this in test logs:
```
Promise returned by `expect(actual).rejects.toThrow()` was not awaited. Vitest currently auto-awaits hanging assertions at the end of the test, but this will cause the test to fail in Vitest 3. Please remember to await the assertion.
```

# Changes

1. Add `await` in front of `expects` which have to await a promise
because they use `.rejects.` or `.resolves.`.

# Tests

Tests no longer complain in the logs.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary