Update vitest from 2.1.9 to 3.0.5 (#6401)
# Motivation

To update Svelte to v5, we need to update `vitest`.
This PR updates just `vitest` and `@vitest/coverage-v8`.

One susprising change in Vitest 3, is that `spyOn` now reuses a mock if
a method is already spyed on.
This means that you might think you're creating a new mock which has
been called 0 times, but actually it was already called several times
before.

# Changes

1. Ran `npm i vitest@latest`
2. Ran `npm i @vitest/coverage-v8@latest`
3. Address deprecation warning that "basic" reporter will go away by
using "default" instead in `vitest.config.ts`.
4. Add `spyQueryProposalApi.mockClear();` after creating a new spy to
reset counters since the mock is not a new mock.
5. Increase an `advanceTime` in
`frontend/src/tests/lib/pages/SnsWallet.spec.ts` to make sure the
transition is passed. This was missed from
https://github.com/dfinity/nns-dapp/pull/6398


# Tests

Existing tests pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary