feat(neurons): add script to create neurons (#6363)
# Motivation

To test the pagination for `list_neurons`, introduced in
https://github.com/dfinity/ic-js/pull/833, an account must have more
than 500 neurons. Ideally, this account should be an Internet Identity
to evaluate its behavior in the UI and when generating a report.

The script will create the specified number of neurons, which will be
controlled by the provided `principalId`.

Here’s an example of how to create 100 neurons for a given principal:
```
scripts/create_neurons -c 100 -p <principalId>
```

# Changes

- Added new script.

# Tests

- Tested locally

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary