Update @dfinity/principal, @dfinity/candid and @dfinity/auth-client (#6402)
# Motivation

In https://github.com/dfinity/nns-dapp/pull/6379 we updated
`@dfinity/agent` to 2.3.0.
`@dfinity/principal`, `@dfinity/candid` and `@dfinity/auth-client` are
all from the same repo and keep their versions in sync.
So it might be a good idea to update those to 2.3.0 as well.

One of these dependencies removed a dependency on `@peculiar/webcrypto`
but our tests still depend on it so it is now added as a test
dependency.

# Changes

1. Ran `npm i @dfinity/auth-client@2.3.0`
2. Ran `npm i @dfinity/candid@2.3.0`
3. Ran `npm i @dfinity/principal@2.3.0`
4. Ran `npm i @peculiar/webcrypto -D`

# Tests

Pass

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary