feat: revert "use test safe transitions" (#6405)
# Motivation

Revert #6396 because thanks to a different way of mocking the
transitions for test purposes (without vite), those workaround are not
needed anymore when migrating to Svelte v5.

# Resources

New solution to mock the animation for test purposes with Svelte v5
providing by Testing Library in
https://github.com/testing-library/svelte-testing-library/issues/284#issuecomment-2651620976

# Changes

Revert #6396

# Tests

I tested the revert and new way of mocking with Svelte v5 in #6404. The
tests did not throw the keyword `TypeError: Cannot set properties of
undefined (setting 'onfinish')`.

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>