Topic filter sorting (#6400)
# Motivation

To improve the user experience, proposal topics need to be displayed in
a sorted order ([the
ticket](https://dfinity.atlassian.net/browse/NNS1-3491)). This applies
to the topic filter and the follow neuron topic list.

In this PR, we address the topic filter sorting. The follow neurons
modal will be updated in a separate PR.

**Expected order:**
1. Governance.
2. SNS & Neurons' Fund.
3. All Except Governance, and SNS & Neurons' Fund. (not available in
Topic filter)
4. Rest topics (ExchangeRate included) sorted alphabetically.

# Changes

- Add sortNnsTopics util.
- Use it to sort topic filter.

# Tests

- Updated.

# Todos

- [x] Add entry to changelog (if necessary).