fix(tests): resolve flaky test due timing in the CI (#6412)
# Motivation

In #6391, the merge queue failed due to a flaky test. The investigation
revealed that the issue occurs when scrolling down to the button while
the skeleton is present. At other times, the component loads, and the
button moves out of the viewport, resulting in this strange behavior.

# Changes

- It waits for the component to render and for the skeleton to disappear
before clicking.

# Tests

- Should pass as before.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.