Remove this.waitForAbsent() from page objects (#6417)
# Motivation

We have several page object function that wait for the component to be
absent by calling `await this.waitForAbsent()` or `await
this.waitForClosed()`.
But a component can't control its own presence; it must be controlled by
the parent component.
So these functions don't work when the component is tested directly but
only if the parent is tested.
This makes some page object method less useful than they could be.
The right thing to do it to move the `waitForAbsent()` call to the page
object of the parent component.

# Changes

1. Remove `this.waitForAbsent()` calls from page objects and instead
wait for the component to be absent on the calling side.

# Tests

1. All the changed code is test-only code.
2. I made sure I didn't miss any callers by renaming the affected method
and then renaming them back after everything compiled.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary