Script to mark PR as ready after tests pass (#6427)
# Motivation

I like to make sure tests pass before marking a PR as ready to review.
But this means I have to remember to get back to the PR to check if the
tests passed and then mark it ready for review.

# Changes

Add `scripts/github-mark-ready` which watches a specified PR and marks
it as ready for review after the tests have passed.
Optionally marking it as "merge when ready".

# Tests

Used manually for 2 separate PRs.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary