chore(linting): update ESLint to latest  version (#6430)
# Motivation

We want to update ESLint to its latest version, as it is no longer
supported. This follows up on #6391, and it is important because
ESLint@8 is outdated.

![Screenshot 2025-02-15 at 11 30
43](https://github.com/user-attachments/assets/c6228b1b-6f13-4980-8348-82341828b486)

To minimize the necessary changes due to the differences in how some
rules apply now compared to before, I've overwritten certain rules. This
will be addressed in a follow-up.

# Changes

- Upgrades ESLint and relevant packages to the latest version.  
- Installs new required packages.  
- Removes packages that are no longer needed.  

# Tests

- Linting should pass as it did previously.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary