feat(metrics): service to initialize plausible analytics (#6454)
# Motivation

As announced in this [forum
post](https://forum.dfinity.org/t/internet-identity-and-nns-dapp-analytics/40859),
we aim to integrate an online analytics solution to monitor errors and
aggregate usage data for the NNS dapp.

Here are some key points about Plausible:
-  No cookies or personal data collection.
-  Aggregates anonymized metrics.
-  Data sovereignty.
-  Hosted in the EU, ensuring strong privacy protections.
-  Complies with GDPR, PECR, and CCPA.
-  Publicly accessible analytics dashboard for community oversight.
-  Minimal data collection focused on critical metrics.
-  Lightweight.

This PR follows up on #6450 by using this new env. variable to expose a
service that allows initializing the tool.

# Changes

- New service to initialize the tool and return a list of cleanup
functions.

# Tests

- Added unit test for the service
- Manually tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network/)
with #6445

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet.