Fix usage of runResolvedPromises() in ImportTokenModal.spec.ts (#6455)
# Motivation

The purpose of `runResolvedPromises()` is to make sure things that
should have already happened, have happened.
Adding more `await runResolvedPromises()` to a test should never cause
it to fail, because the things that it causes to happen could have
happened anyway, if the framework changes what is sync/async etc.

So the following pattern is wrong and brittle and should be avoided
```
expect state A

await runResolvedPromises()

expect state B
```
If we want something like that, we should always explicitly make sure
that whatever causes state B doesn't happen at first and happens later.

The wrong pattern was used `ImportTokenModal.spec.ts`.
This caused the test to fail when trying to upgrade to Svelte 5.


# Changes

1. Change tests that use the wrong pattern described above, to
explicitly block by having API functions return unresolved promises and
then only resolving those promises before the new state is expected.

# Tests

1. The test still passes.
2. The test now also passes in
https://github.com/dfinity/nns-dapp/pull/6020

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary