feat(metrics): integrate plausible analytics  (#6445)
# Motivation

As announced in this [forum
post](https://forum.dfinity.org/t/internet-identity-and-nns-dapp-analytics/40859),
we aim to integrate an online analytics solution to monitor errors and
aggregate usage data for the NNS dapp.

Here are some key points about Plausible:
-  No cookies or personal data collection.
-  Aggregates anonymized metrics.
-  Data sovereignty.
-  Hosted in the EU, ensuring strong privacy protections.
-  Complies with GDPR, PECR, and CCPA.
-  Publicly accessible analytics dashboard for community oversight.
-  Minimal data collection focused on critical metrics.
-  Lightweight.

The data collected on the non-mainnet will be displayed
[here](https://plausible.io/test.nns.ic0.app/).

The data collected on the mainnet will be displayed
[here](https://plausible.io/nns.ic0.app/) once the proposal with these
changes is approved.

This PR follows up on #6454 by using the service after the main layout
has loaded.

![Screenshot 2025-02-19 at 08 26
44](https://github.com/user-attachments/assets/f0434b7d-c01a-4f87-98e2-83ace102224d)

# Changes

- Loads Plausible when the main layout mounts.
- Adds `https://plausible.io/api/event` to list of csp exceptions.

# Tests

- Added unit test to check that the tool is loaded once.
- Manually tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network/)
with #6445

# Todos

- [x] Add entry to changelog (if necessary).